Add files.py
Set up the framework for the File class and its subclasses.
