+++ docs/Linux File Types.md
@@ -1,0 +1,161 @@
+## Conversation on Linux File Types and Importing Modules Dynamically
+
+### **User's Question:**
+To avoid circular imports, I've been having to import all the subclasses of the `File` class in the methods that use them, like this:
+
+```python
+    def check_type(self):
+        from directory import Directory
+        from jsonfile import JSONFile
+        from non_exist_file import NonExistFile
+        from textfile import TextFile
+        return type(self)
+```
+
+This just gets cumbersome. In addition, if the module is being imported as part of a package (`__package__ == True`), there needs to be a period before the name of each module, which makes the whole process twice as cumbersome. What's the best way to get around this? I'd like to be able to just import another module that imports the subclasses, for instance. Or execute a file even.
+
+---
+
+### **Solution: Using a Central Module for Imports**
+Instead of importing each subclass separately inside the method, you can create a central module (e.g., `file_types.py`) that imports all subclasses. Then, your `File` class can import from that module.
+
+#### **Example Structure:**
+```
+your_package/
+│── __init__.py
+│── file.py
+│── file_types.py
+│── directory.py
+│── jsonfile.py
+│── non_exist_file.py
+│── textfile.py
+```
+
+#### **`file_types.py`:**
+```python
+from .directory import Directory
+from .jsonfile import JSONFile
+from .non_exist_file import NonExistFile
+from .textfile import TextFile
+
+# Optional: Collect them in a dictionary or list
+file_classes = {cls.__name__: cls for cls in (Directory, JSONFile, NonExistFile, TextFile)}
+```
+
+#### **`file.py`:**
+```python
+from .file_types import *  # This will make all subclasses available
+```
+
+Now, inside `check_type`, you don't need to import them again.
+
+---
+
+### **Solution: Using `importlib` for Dynamic Importing**
+If your subclasses always reside in a known directory, you can dynamically import them when needed.
+
+```python
+import importlib
+
+def check_type(self):
+    module_names = ["directory", "jsonfile", "non_exist_file", "textfile"]
+    for name in module_names:
+        importlib.import_module("." + name, package=__package__)
+    return type(self)
+```
+This avoids hardcoding multiple imports throughout your code.
+
+---
+
+### **Importing a Class Dynamically from a String**
+You can use `importlib.import_module` along with `getattr` to dynamically import both the module and a specific class from it.
+
+```python
+import importlib
+
+def import_class(module_name, class_name):
+    module = importlib.import_module(module_name)  # Import the module
+    return getattr(module, class_name)  # Retrieve the class from the module
+```
+
+#### **Example Usage:**
+```python
+DirectoryClass = import_class("your_package.directory", "Directory")
+JSONFileClass = import_class("your_package.jsonfile", "JSONFile")
+
+# Create instances dynamically
+dir_instance = DirectoryClass()
+json_instance = JSONFileClass()
+```
+
+If you're working inside a package, you might need relative imports:
+```python
+module = importlib.import_module("." + module_name, package=__package__)
+```
+
+---
+
+### **Linux File Types Beyond Text and Binary**
+Linux has several file types beyond just text and binary:
+
+#### **1. Regular Files**
+- **Text Files**: Plain text, source code, configuration files.
+- **Binary Files**: Executables, images, audio, etc.
+
+#### **2. Special Files**
+- **Symbolic Links (Symlinks)**: Shortcuts pointing to other files.
+- **Device Files**:
+  - **Character Devices** (`/dev/null`, `/dev/tty`): Read/write character by character.
+  - **Block Devices** (`/dev/sda1`): Read/write in blocks (disks, USB drives).
+- **Named Pipes (FIFOs)**: For inter-process communication.
+- **Sockets**: Used for network communication or IPC (e.g., `/var/run/docker.sock`).
+
+#### **3. Directories**
+- Special file type that contains other files.
+
+#### **4. Special System Files**
+- **Proc Files (`/proc/*`)**: Virtual files representing system and process info.
+- **SysFS Files (`/sys/*`)**: Exposes kernel and hardware details.
+
+---
+
+### **Checking If a File is a Socket or a Proc File**
+
+#### **Using `os` and `stat` in Python**
+```python
+import os
+import stat
+
+def check_file_type(file_path):
+    if not os.path.exists(file_path):
+        return "File does not exist"
+
+    mode = os.stat(file_path).st_mode
+
+    if stat.S_ISSOCK(mode):
+        return "Socket File"
+    elif file_path.startswith("/proc/"):
+        return "Proc File"
+    else:
+        return "Other File Type"
+
+# Example Usage
+print(check_file_type("/var/run/docker.sock"))  # Likely a Socket
+print(check_file_type("/proc/cpuinfo"))         # Proc File
+```
+
+#### **Using the `file` Command (Shell)**
+```sh
+file /var/run/docker.sock
+file /proc/cpuinfo
+```
+
+---
+
+### **Summary**
+- You can avoid circular imports by using a **central import module** (`file_types.py`).
+- **Dynamic class importing** can be done using `importlib.import_module` and `getattr`.
+- Linux has several file types beyond text and binary, including **symbolic links, device files, sockets, and proc files**.
+- You can determine if a file is a **socket or a proc file** using `os.stat()` and `stat.S_ISSOCK()`.
+
+Would you like any refinements or additional details?
