--- hw/files/__init__.py
@@ -1,3 +1,0 @@
-__doc__ = """files.py
-
-"""
