--- hw/files/textfile.py
@@ -1,9 +1,0 @@
-if __package__:
-    from .files import File
-else:
-    from files import File
-
-class TextFile(File):
-    def to_upper(self):
-        return self.read().upper()
-
